+++ src/main/java/com/github/johnnyjayjay/discord/commandapi/CommandListener.java
@@ -26,9 +26,9 @@ class CommandListener extends ListenerAdapter {
 
     @Override
     public void onGuildMessageReceived(GuildMessageReceivedEvent event) {
-        String raw = event.getMessage().getContentRaw();
-        String prefix = settings.getPrefix(event.getGuild().getIdLong());
-        if (!event.getAuthor().isBot()) {
+        if (!settings.getBlacklistedChannels().contains(event.getChannel().getIdLong()) && (!event.getAuthor().isBot() || settings.botsMayExecute())) {
+            String raw = event.getMessage().getContentRaw();
+            String prefix = settings.getPrefix(event.getGuild().getIdLong());
             if (raw.startsWith(prefix)) {
                 long timestamp = System.currentTimeMillis();
                 long userId = event.getAuthor().getIdLong();
@@ -47,13 +47,12 @@ class CommandListener extends ListenerAdapter {
     }
 
     private void sendInfo(Member member, TextChannel channel, String prefix, String[] args) {
-        var builder = new EmbedBuilder().setTitle("Help");
+        EmbedBuilder builder = new EmbedBuilder().setTitle("Help");
         if (args.length == 0) {
-            String helpLabels = format("[%s]", join("|", settings.getHelpLabels().toArray(new String[settings.getHelpLabels().size()])));
+            String helpLabels = format("[%s]", join("|", settings.getHelpLabels()));
             builder.appendDescription(format("To learn more about a specific command, just call `%s%s <label>`.\n", prefix, helpLabels))
                     .appendDescription("The following commands are currently available:\n");
-            var commandLabels = settings.getCommands().keySet().toArray(new String[settings.getCommands().keySet().size()]);
-            builder.addField("Commands", format("```\n%s%s```", prefix, join(format(", %s", prefix), commandLabels)), false);
+            builder.addField("Commands", format("```\n%s%s```", prefix, join(format(", %s", prefix), settings.getCommands().keySet())), false);
             channel.sendMessage(builder.build()).queue();
         } else if (args.length == 1 && settings.getCommands().containsKey(args[0])) {
             builder.appendDescription(format("**Command Info for:** `%s`\n\n", args[0]))